Loadng to levels
Added scripts to load to the next level on completion of the current level, carrying over the Player and Main Camera.

There are some issues on Level 1 when restarting the level on Player Death due to the objects existing in the scene at before playing.

Added code to instantiate a Player and a Main Camera if they do not exist in the scene already.  (Left Player and Main Camera in Level 1 becuase removing them caused errors with purple square enemy).

Can move between levels manually by pressing "n" to go forward, and "b" to go back. (There is an error going back to level 1 because of the existing objects in the scene).