Adjusted Camera Script
Reworked Camera script to enable clamping the camera view.  Set initial values of clamping for level 1.

Created BossArena for Level 1 and implemented camera functions that will only view the boss arena once entered. BossArena left and right edges enable on trigger so that the player is not able to leave the BossArena once entered.